Time: 10 ms (82.77%) | Memory: 47.2 MB (54.22%) - LeetSync

+++ 462-minimum-moves-to-equal-array-elements-ii/minimum-moves-to-equal-array-elements-ii.java
@@ -2,10 +2,10 @@ class Solution {
     public int minMoves2(int[] nums) {
         Arrays.sort(nums);
         int median = nums[nums.length/2];
-        int ans = 0;
+        long ans = 0;
         for(int i : nums){
             ans += Math.abs(median-i);
         }
-        return ans;
+        return (int) ans;
     }
 }
